displayed recipe name and image in the RecyclerView
