--- app/src/main/java/com/example/inventory/myRecipe.kt
@@ -1,3 +1,0 @@
-package com.example.inventory
-
-data class myRecipe(var titleImage : Int, var recipeName : String)
